Adição da tela que separa paciente e psicólogo
