+++ src/main/java/com/psico/apoia/app/controller/PrimeiroAcessoViewController.java
@@ -1,18 +1,12 @@
 package com.psico.apoia.app.controller;
 
 import com.psico.apoia.app.common.Usuario;
-import com.psico.apoia.app.service.ICadastroService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class PrimeiroAcessoViewController {
-
-    @Autowired
-    private ICadastroService cadastroService;
 
     @GetMapping("/carregar-primeiro-acesso")
     public String carregarPrimeiroAcesso(Model model) {
@@ -20,15 +14,14 @@ public class PrimeiroAcessoViewController {
         return "primeiro_acesso";
     }
 
-    @PostMapping("/cadastrar")
-    public String cadastrar(Model model, Usuario usuario) {
-        try {
-            cadastroService.cadastrar(usuario);
-        } catch(Exception e) {
-            model.addAttribute("mensagemErro", e.getMessage());
-            return "primeiro_acesso";
-        }
-        model.addAttribute("mensagemSucesso", "Cadastro realizado com sucesso!");
-        return "login";
+    @GetMapping("/primeiro-acesso-paciente")
+    public String carregarPaginaPaciente(Model model) {
+        model.addAttribute("usuario", new Usuario());
+        return "primeiro_acesso_paciente";
+    }
+    @GetMapping("/primeiro-acesso-psicologo")
+    public String carregarPaginaPsicologo(Model model) {
+        model.addAttribute("usuario", new Usuario());
+        return "primeiro_acesso_psicologo";
     }
 }
