+++ src/main/resources/static/css/primeiro_acesso.css
@@ -1,68 +1,68 @@
 body {
-  width: 100%;
-  height: 100%;
-  margin: 0 auto;
-  background-color: #EB980A;
+    width: 100%;
+    height: 100%;
+    margin: 0 auto;
+    background-color: #EB980A;
 }
 
 main {
-  margin: 0 auto;
-  width: 1580px;
-  position: relative;
+    margin: 0 auto;
+    width: 1580px;
+    position: relative;
 }
 
 
 
 @media only screen and (max-width: 768px) {
-  .menu ul li {
-    margin-right: 10px;
-    margin-bottom: 10px;
-  }
+    .menu ul li {
+        margin-right: 10px;
+        margin-bottom: 10px;
+    }
 }
 
 .header-3 {
-  margin-top: 5px;
-  width: 100%;
-  height: auto;
+    margin-top: 5px;
+    width: 100%;
+    height: auto;
 }
 
 .form-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
 }
 
 .form-cadastro {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: #FFE9C3;
-  border-radius: 10px;
-  padding: 20px;
-  width: 50%;
-  margin: 0 auto;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    background-color: #FFE9C3;
+    border-radius: 10px;
+    padding: 20px;
+    width: 50%;
+    margin: 0 auto;
 }
 
 @media screen and (max-width: 768px) {
-  .form-cadastro {
-    max-width: 90%;
-  }
+    .form-cadastro {
+        max-width: 90%;
+    }
 }
 
 @media screen and (max-width: 768px) {
-  .form-cadastro {
-    flex-direction: column;
-  }
+    .form-cadastro {
+        flex-direction: column;
+    }
 
-  .form-group {
-    width: 100%;
-  }
+    .form-group {
+        width: 100%;
+    }
 }
 
 .form-group {
     color: #f08700;
-    font-family: Arial, Helvetica, sans-serif
+    font-family: Arial, Helvetica, sans-serif;
     margin-top: 5px;
     text-align: left;
     flex: 1;
@@ -72,7 +72,7 @@ main {
 
 .form-group-1 {
     color: #f08700;
-    font-family: Arial, Helvetica, sans-serif
+    font-family: Arial, Helvetica, sans-serif;
     margin-top: 10px;
     display: flex;
     flex-direction: row;
@@ -82,39 +82,68 @@ main {
 }
 
 .form-group-1-inputs {
-  width: calc(33.33% - 10px);
-  margin-bottom: 10px;
+    width: calc(33.33% - 10px);
+    margin-bottom: 10px;
 }
 
 
 label {
-  font-family: Arial, Helvetica, sans-serif;
-  font-weight: bold;
-  display: block;
-  margin-bottom: 5px;
-  white-space: nowrap;
+    font-family: Arial, Helvetica, sans-serif;
+    font-weight: bold;
+    display: block;
+    margin-bottom: 5px;
+    white-space: nowrap;
 }
 
 input {
-  border: 2px solid #f08700;
-  border-radius: 5px;
-  padding: 6px;
-   width: 100%;
-   height: auto;
+    border: 2px solid #f08700;
+    border-radius: 5px;
+    padding: 6px;
+    width: 100%;
+    height: auto;
+}
+
+.choice-button-group {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    position: relative;
+    height: 100vh;
+}
+
+.choice-button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: #FFE9C3;
+    color: #F08700;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 30px;
+    margin-right: 10px;
+    margin-bottom: 30px;
+    cursor: pointer;
+    position: relative;
+    text-decoration: none;
+    width: 33%;
+    box-sizing: border-box;
+    font-family: Arial, Helvetica, sans-serif;
+    font-size: 16px;
 }
 
 .button-group button {
-  background-color: #f08700;
-  color: #fff;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 5px;
-  margin-right: 10px;
-  cursor: pointer;
+    background-color: #f08700;
+    color: #fff;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 20px;
+    margin-right: 10px;
+    cursor: pointer;
 }
 
 .form-group button {
-  margin-bottom: 0;
+    margin-bottom: 0;
 }
