+++ src/main/resources/templates/primeiro_acesso_psicologo.html
@@ -1,0 +1,80 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <link th:href="@{/css/primeiro_acesso.css}" rel="stylesheet"/>
+    <link th:href="@{/css/common.css}" rel="stylesheet"/>
+    <link th:href="@{/css/toastr.min.css}" rel="stylesheet"/>
+    <script type="text/javascript" th:src="@{/js/jquery-3.7.0.js}"></script>
+    <script type="text/javascript" th:src="@{/js/toastr.min.js}"></script>
+    <script type="text/javascript" th:src="@{/js/common.js}"></script>
+    <title>Psico Apoia</title>
+</head>
+<body>
+
+<div th:replace="~{cabecalho :: cabecalho}"></div>
+
+<div class="header-3">
+    <div class="form-container">
+        <div class="form-cadastro">
+            <form action="#" th:action="@{/cadastrar}" th:object="${usuario}" method="post">
+                <div class="form-group-1">
+                    <label for="usuario">Usuário:</label>
+                    <input class="form-group-1-inputs"  id="usuario" name="usuario" type="text" th:field="*{usuario}"/>
+                    <label for="senha">Senha:</label>
+                    <input class="form-group-1-inputs"  id="senha" name="senha" type="password" th:field="*{senha}"/>
+                    <label for="senha2">Confirmação Senha:</label>
+                    <input class="form-group-1-inputs"  id="senha2" name="senha2" type="password" th:field="*{senhaConfirmacao}"/>
+                </div>
+                <div class="form-group">
+                    <label for="nome_completo">Nome completo:</label>
+                    <input id="nome_completo" name="nome_completo" type="text" th:field="*{nome}"/>
+                    <label for="crp">CRP</label>
+                    <input class="form-group-1-inputs" is="crp" name="crp" type="text" th:field="*{numeroCrp}">
+                </div>
+                <div class="form-group-1">
+                    <label for="data_nascimento">Data de nascimento:</label>
+                    <input class="form-group-1-inputs" id="data_nascimento" name="data_nascimento" type="date" th:field="*{dataNascimento}"/>
+                    <label for="cpf">CPF:</label>
+                    <input class="form-group-1-inputs" id="cpf" name="cpf" type="text" th:field="*{numeroCpf}"/>
+                    <label for="rg">RG:</label>
+                    <input class="form-group-1-inputs" id="rg" name="rg" type="text" th:field="*{numeroRg}"/>
+                </div>
+                <div class="form-group">
+                    <label for="email">E-mail:</label>
+                    <input id="email" name="email" type="email" th:field="*{email}"/>
+                </div>
+                <div class="form-group">
+                    <label for="telefone">Telefone com DDD:</label>
+                    <input id="telefone" name="telefone" type="text" th:field="*{telefone}"/>
+                </div>
+                <div class="form-group-1">
+                    <label for="cep">CEP:</label>
+                    <input class="form-group-1-inputs"  id="cep" name="cep" type="text" th:field="*{endereco.cep}"/>
+                    <label for="uf">UF:</label>
+                    <input class="form-group-1-inputs"  id="uf" name="uf" type="text" th:field="*{endereco.uf}"/>
+                    <label for="cidade">Cidade:</label>
+                    <input class="form-group-1-inputs"  id="cidade" name="cidade" type="text" th:field="*{endereco.cidade}"/>
+                    <label for="bairro">Bairro:</label>
+                    <input class="form-group-1-inputs"  id="bairro" name="bairro" type="text" th:field="*{endereco.bairro}"/>
+                </div>
+                <div class="form-group-1">
+                    <label for="logradouro">Logradouro:</label>
+                    <input class="form-group-1-inputs"  id="logradouro" name="logradouro" type="text" th:field="*{endereco.logradouro}"/>
+                    <label for="numero">Número:</label>
+                    <input class="form-group-1-inputs"  id="numero" name="numero" type="text" th:field="*{endereco.numeroResidencia}"/>
+                    <label for="complemento">Complemento:</label>
+                    <input class="form-group-1-inputs" id="complemento" name="complemento" type="text" th:field="*{endereco.complemento}"/>
+                </div>
+                <div class="form-group button-group">
+                    <button type="reset">Limpar</button>
+                    <button class="submit-btn" type="submit">Salvar</button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+
+<div th:replace="~{rodape :: rodape}"></div>
+
+</body>
+</html>
